+++ app/Http/Controllers/PhaseTempsPriseController.php
@@ -19,19 +19,24 @@ class PhaseTempsPriseController extends Controller
     public function __construct(PhaseTempsPriseServiceInterface $service)
     {
         $this->service = $service;
-        $this->middleware(function ($request, $next) {
-            if (Gate::allows('isSuperAdmin')) {
-                return $next($request);
-            }
+        // $this->middleware(function ($request, $next) {
+        //     if (Gate::allows('isSuperAdmin')) {
+        //         return $next($request);
+        //     }
 
-            return $this->responseError('Unauthorized', 403);
-        })->except('index');
+        //     return $this->responseError('Unauthorized', 403);
+        // });
     }
 
     public function index(): JsonResponse
     {
         try {
-            $data = $this->service->all();
+            if (Gate::allows('isSuperAdmin') || Gate::allows('isAdmin') || Gate::allows('isUser')) {
+
+                $data = $this->service->all();
+            } else {
+                return $this->responseError('Unauthorized', 403);
+            }
         } catch (Exception $e) {
             return $this->responseError($e->getMessage());
         }
@@ -43,7 +48,7 @@ class PhaseTempsPriseController extends Controller
         $payload = PhaseTempsPriseDTO::fromAdd($request->all());
 
         try {
-            if (Gate::allows('isSuperAdmin') || Gate::allows('Admin')) {
+            if (Gate::allows('isSuperAdmin') || Gate::allows('Admin') || Gate::allows('isUser')) {
                 $data = $this->service->store($payload);
             } else {
                 return $this->responseError('Unauthorized', 403);
@@ -58,7 +63,7 @@ class PhaseTempsPriseController extends Controller
     public function edit(PhaseTempsPriseRequest $request, int $id)
     {
         try {
-            if (Gate::allows('isSuperAdmin') || Gate::allows('Admin')) {
+            if (Gate::allows('isSuperAdmin') || Gate::allows('isAdmin')) {
                 $data = $this->service->edit($request->all(), $id);
             } else {
                 return $this->responseError('Unauthorized', 403);
@@ -73,7 +78,11 @@ class PhaseTempsPriseController extends Controller
     public function destroy(int $id): JsonResponse
     {
         try {
-            $this->service->destroy($id);
+            if (Gate::allows('isSuperAdmin')) {
+                $this->service->destroy($id);
+            } else {
+                return $this->responseError('Unauthorized', 403);
+            }
         } catch (Exception $e) {
             return $this->responseError($e->getMessage());
         }
